remove CSL headers from public API
