+++ modules/dnn/src/layers/fully_connected_layer.cpp
@@ -424,18 +424,19 @@ public:
 
 #ifdef HAVE_CUDA
     Ptr<BackendNode> initCUDA(
-        csl::Stream stream,
-        csl::cublas::Handle cublas_handle,
-        csl::cudnn::Handle cudnn_handle,
-        const std::vector<Ptr<BackendWrapper>>& inputs
+        void *context_,
+        const std::vector<Ptr<BackendWrapper>>& inputs,
+        const std::vector<Ptr<BackendWrapper>>& outputs
     ) override
     {
+        auto context = reinterpret_cast<csl::CSLContext*>(context_);
+
         auto input_wrapper = inputs[0].dynamicCast<CUDABackendWrapper>();
 
         auto flatten_start_axis = clamp(axis, input_wrapper->getRank());
 
         auto biasMat_ = bias ? biasMat : Mat();
-        return make_cuda_node<cuda4dnn::InnerProductOp>(preferableTarget, std::move(stream), std::move(cublas_handle), flatten_start_axis, weightsMat, biasMat_);
+        return make_cuda_node<cuda4dnn::InnerProductOp>(preferableTarget, std::move(context->stream), std::move(context->cublas_handle), flatten_start_axis, weightsMat, biasMat_);
     }
 #endif
